+++ package.json
@@ -31,9 +31,12 @@
     "expo-linking": "~6.3.1",
     "expo-media-library": "~16.0.3",
     "expo-permissions": "^14.4.0",
+    "expo-secure-store": "~13.0.1",
+    "expo-sharing": "~12.0.1",
     "expo-sqlite": "~14.0.3",
     "expo-status-bar": "~1.12.1",
     "expo-web-browser": "~13.0.3",
+    "moment": "^2.30.1",
     "react": "18.2.0",
     "react-native": "0.74.2",
     "react-native-blob-util": "^0.19.9",
@@ -52,12 +55,11 @@
     "react-native-reanimated": "~3.10.1",
     "react-native-safe-area-context": "4.10.1",
     "react-native-screens": "3.31.1",
-    "react-native-svg": "15.2.0",
+    "react-native-svg": "^15.2.0",
+    "react-native-swipe-gestures": "^1.0.5",
     "react-native-vector-icons": "^10.0.3",
     "react-native-webview": "13.8.6",
-    "realm": "^12.9.0",
-    "expo-secure-store": "~13.0.1",
-    "expo-sharing": "~12.0.1"
+    "realm": "^12.9.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
