+++ src/screens/Home/Dashboard.js
@@ -19,6 +19,7 @@ import {
   Alert,
   Dimensions,
 } from "react-native";
+import ChartsCard from "./ChartsCard";
 const UserProfileCard = ({ user, onPress }) => {
   if (!user) {
     return (
@@ -64,158 +65,6 @@ const fetchDataByEmail = async (email) => {
   }
 };
 
-const ChartsCard = ({ userId }) => {
-  const [modalVisible, setModalVisible] = useState(false);
-  const [weight, setWeight] = useState("");
-  const [height, setHeight] = useState("");
-  const [notes, setNotes] = useState("");
-  const [userData, setUserData] = useState([]);
-
-  const handleAddButtonPress = () => {
-    setModalVisible(true);
-  };
-
-  const handleSubmit = async () => {
-    if (!weight || !height) {
-      Alert.alert("Error", "Please enter both weight and height.");
-      return;
-    }
-
-    Alert.alert(
-      "Confirm Submission",
-      "Are you sure you want to add this data?",
-      [
-        {
-          text: "Cancel",
-          style: "cancel",
-        },
-        {
-          text: "OK",
-          onPress: async () => {
-            const data = { weight, height, notes, user: userId, date: new Date() };
-            try {
-              await axios.post('http://192.168.205.43:33000/api/healthData', data);
-              setModalVisible(false);
-              fetchUserData();  // Fetch updated user data after submission
-            } catch (error) {
-              console.error("Error saving health data:", error);
-            }
-          },
-        },
-      ]
-    );
-  };
-
-  const handleCancel = () => {
-    setModalVisible(false);
-  };
-
-  const fetchUserData = async () => {
-    try {
-      const response = await axios.get(`http://192.168.205.43:33000/api/healthData/userId/${userId}`);
-      setUserData(response.data);
-    } catch (error) {
-      console.error("Error fetching user data:", error);
-      setUserData([]); // Set an empty array to ensure chart renders without data
-    }
-  };
-
-  useEffect(() => {
-    if (userId) {
-      fetchUserData();
-    }
-  }, [userId]);
-
-  const calculateBMI = (weight, height) => {
-    if (!weight || !height) {
-      return 0; // Return 0 if weight or height is not provided
-    }
-    const heightInMeters = height / 100;
-    return parseFloat((weight / (heightInMeters * heightInMeters)).toFixed(2));
-  };
-
-  const chartData = userData.map(data => {
-    if (data.weight && data.height) {
-      return calculateBMI(data.weight, data.height);
-    }
-    return 0; // Default to 0 if weight or height is missing
-  });
-
-  const defaultData = {
-    labels: userData.length > 0 ? userData.map(data => new Date(data.date).toLocaleString('default', { month: 'short' })) : ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'],
-    datasets: [{
-      data: chartData.length > 0 ? chartData : [0, 0, 0, 0, 0, 0]
-    }],
-  };
-
-  return (
-    <View>
-      <Text style={styles.textHeader}>Your BMI</Text>
-      <TouchableOpacity style={styles.addButton} onPress={handleAddButtonPress}>
-        <Icon name="add-circle-outline" size={30} color="#000" />
-      </TouchableOpacity>
-      <LineChart
-        data={defaultData}
-        width={Dimensions.get("window").width}
-        height={220}
-        yAxisLabel=""
-        yAxisSuffix=""
-        yAxisInterval={1}
-        chartConfig={{
-          backgroundColor: "#ffd600",
-          backgroundGradientFrom: "#ff6d00",
-          backgroundGradientTo: "#ffab00",
-          decimalPlaces: 2,
-          color: (opacity = 1) => `rgba(255, 255, 255, ${opacity})`,
-          labelColor: (opacity = 1) => `rgba(255, 255, 255, ${opacity})`,
-          style: { borderRadius: 16 },
-          propsForDots: {
-            r: "6",
-            strokeWidth: "2",
-            stroke: "#ffa726",
-          },
-        }}
-        bezier
-        fromZero={true} // Ensure Y-axis starts from 0
-        style={{ margin: 10, borderRadius: 16 }}
-      />
-      <Modal
-        animationType="slide"
-        transparent={true}
-        visible={modalVisible}
-        onRequestClose={() => setModalVisible(false)}
-      >
-        <View style={styles.modalView}>
-          <Text style={styles.modalText}>Add Your Health Data</Text>
-          <TextInput
-            style={styles.input}
-            placeholder="Weight (kg)"
-            keyboardType="numeric"
-            value={weight}
-            onChangeText={setWeight}
-          />
-          <TextInput
-            style={styles.input}
-            placeholder="Height (cm)"
-            keyboardType="numeric"
-            value={height}
-            onChangeText={setHeight}
-          />
-          <TextInput
-            style={styles.input}
-            placeholder="Notes"
-            value={notes}
-            onChangeText={setNotes}
-          />
-          <View style={styles.buttonContainer}>
-            <Button title="Submit" onPress={handleSubmit} />
-            <Button title="Cancel" onPress={handleCancel} color="#FF6347" />
-          </View>
-        </View>
-      </Modal>
-    </View>
-  );
-};
 
 const formatDate = (dateString) => {
   const parsedDate = Date.parse(dateString);
@@ -245,13 +94,11 @@ const ListCard = ({ documents, user, navigation }) => {
       style={styles.item}
       onPress={() => openDocument(item, navigation)}
     >
-      <Icon name="document-outline" size={40} color="#000" marginLeft={20} />
-      <View>
+      <Icon name="document-outline" size={50} color="#000" marginLeft={10} />
+      <View style={styles.itemContainer}>
         <Text style={styles.title}>{item.fileName}</Text>
         <Text style={styles.titledescription}>{item.description}</Text>
-        <Text style={styles.uploadDate}>
-          Uploaded on: {formatDate(item.date)}
-        </Text>
+        <Text style={styles.uploadDate}>Uploaded on: {formatDate(item.date)}</Text>
       </View>
     </TouchableOpacity>
   );
@@ -394,10 +241,20 @@ const styles = StyleSheet.create({
     justifyContent: "space-between",
     alignItems: "center",
     backgroundColor: "#bbdefb",
-    padding: 20,
-    marginVertical: 8,
+    flexDirection: "row",
+    padding: 10,
+    margin: 10,
     borderRadius: 10,
+    shadowColor: "#000",
+    shadowOffset: { width: 0, height: 2 } ,
+    shadowOpacity: 0.3,
+    shadowRadius: 2,
+    elevation: 5,
+  
   },
+itemContainer:{
+    marginRight:80,
+},
   title: {
     fontSize: 18,
     fontWeight: "bold",
@@ -418,42 +275,7 @@ const styles = StyleSheet.create({
     paddingBottom: 20,
   },
  
-    addButton: {
-      position: 'absolute',
-      right: 10,
-      top: 10,
-      zIndex: 1,
-    },
-    modalView: {
-      margin: 20,
-      backgroundColor: "white",
-      borderRadius: 20,
-      padding: 35,
-      alignItems: "center",
-      shadowColor: "#000",
-      shadowOffset: { width: 0, height: 2 },
-      shadowOpacity: 0.25,
-      shadowRadius: 4,
-      elevation: 5,
-    },
-    modalText: {
-      marginBottom: 15,
-      textAlign: "center",
-      fontSize: 20,
-    },
-    input: {
-      height: 40,
-      borderColor: "gray",
-      borderWidth: 1,
-      marginBottom: 10,
-      width: "100%",
-      padding: 10,
-    },
-    buttonContainer: {
-      flexDirection: "row",
-      justifyContent: "space-between",
-      width: "100%",
-    },
+   
 
   
 });
